refactor: TastingInfo import 루트 @/shared 로 변경 적용 안된거 추가 수정
