+++ src/features/meetings/model/types.ts
@@ -1,4 +1,4 @@
-import { TastingInfo } from '@/entities/tasting-list/model/types';
+import { TastingInfo } from '@/shared';
 
 export interface PostMeetingRequest {
   meetingTitle: string;
